pyproject.toml
[project]
name = "construct-server"
version = "0.1.0"
description = "Construct automation server"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.0",
    "workflow-use",
    "browser-use",
    "openai>=1.0.0",
    "langchain-openai>=0.3.0",
    "playwright>=1.40.0",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
construct-dev = "dev:main"

[tool.setuptools.packages.find]
include = ["server*"]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"
